+++ src/main/java/com/qiashe/controller/FacilityItemsController.java
@@ -2,6 +2,7 @@ package com.qiashe.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.qiashe.Util.LoginUtil;
 import com.qiashe.dao.TableDao;
 import com.qiashe.domain.CanteenItem;
 import com.qiashe.domain.FI_Avg_Result;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
@@ -42,8 +44,15 @@ public class FacilityItemsController {
     }
 
     @RequestMapping(value = "/add")
-    public ModelAndView addFacilityItem(String title, String info, Float positionx, Float positiony, Float price, String time) {
+    public ModelAndView addFacilityItem(String title, String info, Double positionx, Double positiony, Float price, String time, HttpServletRequest request) {
         ModelAndView mv = new ModelAndView("message");
+        if(!LoginUtil.validateLogon(request)){
+            mv.setViewName("message");
+            mv.addObject("tips","未登录");
+            mv.addObject("before","<a href='login.jsp'>登录</a>");
+            mv.addObject("next","");
+            return mv;
+        }
         if (1 == 0) {
             mv.addObject("tips", "");
             mv.addObject("before", "<a href='facilityItem/addPage'>返回</a>");
@@ -59,8 +68,15 @@ public class FacilityItemsController {
     }
 
     @RequestMapping(value = "/addPage")
-    public ModelAndView toAddPage(String title, String info, Float positionx, Float positiony, Float price, String time) {
+    public ModelAndView toAddPage(String title, String info, Double positionx, Double positiony, Float price, String time,HttpServletRequest request) {
         ModelAndView mv = new ModelAndView("addFacilityItem");
+        if(!LoginUtil.validateLogon(request)){
+            mv.setViewName("message");
+            mv.addObject("tips","未登录");
+            mv.addObject("before","<a href='login.jsp'>登录</a>");
+            mv.addObject("next","");
+            return mv;
+        }
         if(title!=""&&title!=null)
             mv.addObject("title",title);
         if(info!=""&&info!=null){
@@ -103,15 +119,22 @@ public class FacilityItemsController {
         return json;
     }
     @RequestMapping(value = "/remove")
-    public ModelAndView removeItem(Integer id) {
+    public ModelAndView removeItem(Integer id,HttpServletRequest request) {
         ModelAndView mv = new ModelAndView("message");
+        if(!LoginUtil.validateLogon(request)){
+            mv.setViewName("message");
+            mv.addObject("tips","未登录");
+            mv.addObject("before","<a href='login.jsp'>登录</a>");
+            mv.addObject("next","");
+            return mv;
+        }
         String tip = "删除失败";
         int num = facilityItemsService.RemoveItem(id);
         if (num != 0) {
             tip = "删除成功";
         }
         mv.addObject("tips", tip);
-        mv.addObject("next", "<a href='index.jsp'>返回</a>");
+//        mv.addObject("next", "<a href='index.jsp'>返回</a>");
         return mv;
     }
     @RequestMapping(value = "/getAvgPriceAndMark",produces = "application/json;charset=utf-8")
